HEADLINE-AGENT/requirements.txt
python-dotenv
requests
langchain-community
langchain-core
langchain-openai
langchain-groq
langgraph
torch
inflect
uroman 
transformers 
torch 
torchaudio
outetts
ipython
spitch